Eleven: Fix integer overflow
The exception below occurs when the _id field from the media database is
read as an integer instead of a long.  The call stack above this function
correctly returns long values.

Caused by: java.lang.UnsupportedOperationException: Unknown or unsupported URL: content://media/external/audio/artists/-1755600233/albums
	at android.database.DatabaseUtils.readExceptionFromParcel(DatabaseUtils.java:174)
	at android.database.DatabaseUtils.readExceptionFromParcel(DatabaseUtils.java:142)
	at android.content.ContentProviderProxy.query(ContentProviderNative.java:472)
	at android.content.ContentResolver.query(ContentResolver.java:1183)
	at android.content.ContentResolver.query(ContentResolver.java:1115)
	at android.content.ContentResolver.query(ContentResolver.java:1071)
	at org.lineageos.eleven.loaders.AlbumLoader.makeAlbumCursor(AlbumLoader.java:150)
	at org.lineageos.eleven.loaders.AlbumLoader.loadInBackground(AlbumLoader.java:76)
	at org.lineageos.eleven.loaders.AlbumLoader.loadInBackground(AlbumLoader.java:41)
	at androidx.loader.content.AsyncTaskLoader.onLoadInBackground(AsyncTaskLoader.java:307)
	at androidx.loader.content.AsyncTaskLoader$LoadTask.doInBackground(AsyncTaskLoader.java:60)
	at androidx.loader.content.AsyncTaskLoader$LoadTask.doInBackground(AsyncTaskLoader.java:48)
	at androidx.loader.content.ModernAsyncTask$2.call(ModernAsyncTask.java:141)
	at java.util.concurrent.FutureTask.run(FutureTask.java:266)

+++ src/org/lineageos/eleven/utils/MusicUtils.java
@@ -1025,7 +1025,7 @@ public final class MusicUtils {
             if (cursor != null) {
                 cursor.moveToFirst();
                 if (!cursor.isAfterLast()) {
-                    return cursor.getInt(0);
+                    return cursor.getLong(0);
                 }
             }
         }
